[MOD] Just recording some minor changes for now.

+++ src/java/spssio/por/PORFile.java
@@ -181,11 +181,11 @@ public class PORFile {
     // HEADER
     //========
     
-    public byte[] getSplash() {
+    public int[] getSplash() {
         return header.splash;
     }
     
-    public byte[] getCharset() {
+    public int[] getCharset() {
         return header.charset;
     }
